#37 Full Grown Spell App! (#50)
This introduces the spell search app.  It allows all users to search through a (configurable) compendium for all spells based on provided criteria.  They can search all of them by opening the spell search app from the compendium tab or if they have a spellcasting class on their character sheet, it will provide a link that will open a pre-configured search app that filters to that specific class *and* to the appropriate levels of spell that their class can prepare (or learn, if not in Materia).

It has the following features:
- configurable compendium via settings
- a series of "tri-state" filters:  ignore, include, exclude.  Filters in the same "group" will stack include/exclude with each other.
- filter on name via a text search
- filter on class level
- filter on class spell list
- filter on spell school
- filter on various spell "properties"  (VSM, ritual, concentration;  note that this is all considered one "group")
- filter on cast time  (action, BA, reaction, other)
- enriched spell names that can be dragged onto PC sheets or into chat

This was meant to somewhat mimic the 5e.tools search capabilities in a simplified capacity.  Notable things that are different:
- the results table is *not* dynamic (yet) and columns cannot be clicked for sorting
- "target" ranges for "Self" ranged spells (eg, Spirit Guardians has a 15-foot radius even though the spell range is actually "Self") are not displayed

Of course, features can potentially be added/changed/improved so feel free to provide me feedback!