Release 0.9.0 (#51)
This release includes a number of small fixes as well as the introduction of the Spell Search App.  It allows all users to search through a (configurable) compendium for all spells based on provided criteria.  They can search all of them by opening the spell search app from the compendium tab or if they have a spellcasting class on their character sheet, it will provide a link that will open a pre-configured search app that filters to that specific class *and* to the appropriate levels of spell that their class can prepare (or learn, if not in Materia).

It has the following features:
- configurable compendium via settings
- a series of "tri-state" filters:  ignore, include, exclude.  Filters in the same "group" will stack include/exclude with each other.
- filter on name via a text search
- filter on class level
- filter on class spell list
- filter on spell school
- filter on various spell "properties"  (VSM, ritual, concentration;  note that this is all considered one "group")
- filter on cast time  (action, BA, reaction, other)
- enriched spell names that can be dragged onto PC sheets or into chat

This was meant to somewhat mimic the 5e.tools search capabilities in a simplified capacity.  Notable things that are different:
- the results table is *not* dynamic (yet) and columns cannot be clicked for sorting
- "target" ranges for "Self" ranged spells (eg, Spirit Guardians has a 15-foot radius even though the spell range is actually "Self") are not displayed

Of course, features can potentially be added/changed/improved so feel free to provide me feedback!

In addition, the following minor fixes are included as well:
- spells now all include "spellLists" flags (required to enable spell search app)
- the "item use auto-wireup" feature that would update the items on your PC sheet that used another item on your PC sheet was only working during leveling, now it should work for manually adding things to your PC sheet as well
- Find Familiar summon Keen Senses were flagged as disadvantage instead of advantage on Perception checks
- cleaned up some spell descriptions
- fixed missing range on some spells
- fixed missing V component on Silvery Barbs
- added an effect for advantage on WIS saves to Heroes' Feast
- added missing Summon Fiend spell